+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Fix get_ticket_number when branch name contains numbers
 - Load `.env.local` on top of `.env` in case it exists
 - Add `{{BACKGROUND}}` with "Details in the ticket" by default when a ticket link is used
+- Enable spaces inside placeholders
+  - `{{ BACKGROUND }}`
+- Enable placeholders inside HTML comments in the PR template
+  - `<!-- {{ BACKGROUND }} -->`
 
 ## 0.2 - 2024-08-28